vit_tactile/requirements.txt
torch==2.1.2
torchvision==0.16.2
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.66.0
einops>=0.6.0
pytest>=7.4.0
fast-transformers-pytorch>=0.4.0
timm>=0.9.5
wandb>=0.15.0
tensorboard>=2.14.0 